Adjusts database column type for extra IDs
Changes the `extras_ids` column type in the `#__dnbooking_reservations` table from JSON to LONGTEXT.

This modification ensures compatibility and proper data storage for extra IDs.
Also adds an update sql file to change the column type.
Additionally, it adds a check for holidays or weekends when getting blocked rooms.